Increase size of boot partition in EC2 preset
With the kernel growing over time, a 50MB boot partition can be a problem
when trying to use newer 3.x kernels.

Therefore, increase the size of the boot partition from 50MB to 512MB. This
also has the nice feature of being the same size as the boot partition
generated by the generic preset.